test: refactoring of unit tests: separate helper functions from fixtures in conftest.py
