+++ tests/launcher/test_extension_adequacy_patch.py
@@ -7,7 +7,7 @@ from antarest.core.config import Config, StorageConfig
 from antarest.launcher.extensions.adequacy_patch.extension import (
     AdequacyPatchExtension,
 )
-from tests.conftest import with_db_context
+from tests.helpers import with_db_context
 
 
 @with_db_context
