+++ tests/launcher/test_repository.py
@@ -6,12 +6,12 @@ import pytest
 from sqlalchemy import create_engine
 
 from antarest.core.persistence import Base
-from antarest.core.utils.fastapi_sqlalchemy import db, DBSessionMiddleware
-from antarest.launcher.model import JobResult, JobStatus, JobLog, JobLogType
+from antarest.core.utils.fastapi_sqlalchemy import DBSessionMiddleware, db
+from antarest.launcher.model import JobLog, JobLogType, JobResult, JobStatus
 from antarest.launcher.repository import JobResultRepository
 from antarest.study.model import RawStudy
 from antarest.study.repository import StudyMetadataRepository
-from tests.conftest import with_db_context
+from tests.helpers import with_db_context
 
 
 @pytest.mark.unit_test
