+++ tests/worker/test_simulator_worker.py
@@ -2,18 +2,18 @@ import os
 import platform
 import stat
 from pathlib import Path
-from unittest.mock import Mock, patch, call
+from unittest.mock import Mock, call, patch
 
 import pytest
 
 from antarest.core.config import Config, LauncherConfig, LocalConfig
 from antarest.worker.simulator_worker import (
-    SimulatorWorker,
     GENERATE_KIRSHOFF_CONSTRAINTS_TASK_NAME,
     GENERATE_TIMESERIES_TASK_NAME,
+    SimulatorWorker,
 )
 from antarest.worker.worker import WorkerTaskCommand
-from tests.conftest import with_db_context
+from tests.helpers import with_db_context
 
 
 @with_db_context
